+++ app/assets/javascripts/oxalis/api/api_v2.js
@@ -729,6 +729,7 @@ class UserApi {
     - sortTreesByName
     - sortCommentsAsc
     - sphericalCapRadius
+    - hideTreeRemovalWarning
   *
   * @example
   * const keyboardDelay = api.user.getConfiguration("keyboardDelay");